Fix syncing of broken symlinks as the root
When in preserve symlinks mode, a broken symlink is expected to sync just fine, but this wasn't happening if that symlink was a root.

Change SetRoot so that it returns full details of the root, so that GetEntries doesn't need to be called. This was causing problems as the WalkDir crate fails on broken symlinks as the root, which meant we never even got any details for the root.

Add tests for both preserve and unaware mode.

Add a common entry_details_from_metadata function that is now used from both SetRoot and GetEntries.