Change default features to not include embedded binaries
This makes development easier because you often want to disable embedded binaries to speed up the build, and now you don't need to add --no-default-features every time you run cargo. The downside is that you need to explicitly enable embedded binaries when you do want to, and the README has been updated to explain this.